pyproject.toml
[project]
name = "smarthomeemulator"
version = "0.1.0"
description = ""
authors = [
    {name = "bogdnnx",email = "bogdnnx@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "uvicorn[standart] (>=0.35.0,<0.36.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
